django/testproject/applications/product/serializers.py
from rest_framework import serializers

from applications.product.models import Product


class ProductSerializers(serializers.ModelSerializer):
    class Meta:
        model = Product
        fields = '__all__'
