django/testproject/requirements.txt
django==4.0.6
djangorestframework==3.13
psycopg2-binary==2.9.3